+++ rpc/backend/call_tx_test.go
@@ -286,6 +286,26 @@ func (suite *BackendTestSuite) TestSendRawTransaction() {
 	cosmosTx, _ := ethTx.BuildTx(suite.backend.clientCtx.TxConfig.NewTxBuilder(), "aphoton")
 	txBytes, _ := suite.backend.clientCtx.TxConfig.TxEncoder()(cosmosTx)
 
+	// Build a tx whose fee exceeds the default 1 ether cap (gasPrice * gas > 1e18).
+	highFeeTx := evmtypes.NewTx(
+		suite.backend.chainID, 0, &common.Address{}, big.NewInt(0),
+		100000, new(big.Int).Mul(big.NewInt(1e10), big.NewInt(1e9)), // 10 gwei * 100k gas = 1e6 gwei > 1 ether... actually 1e10 * 1e9 = 1e19 wei = 10 ether
+		nil, nil, nil, nil,
+	)
+	highFeeTx.From = suite.signerAddress
+	_ = highFeeTx.Sign(ethtypes.LatestSignerForChainID(suite.backend.chainID), suite.signer)
+	highFeeRlp, _ := rlp.EncodeToBytes(highFeeTx.AsTransaction())
+
+	// Build a tx signed with a different chain ID.
+	wrongChainID := new(big.Int).Add(suite.backend.chainID, big.NewInt(1))
+	wrongChainTx := evmtypes.NewTx(
+		wrongChainID, 0, &common.Address{}, big.NewInt(0),
+		100000, big.NewInt(1), nil, nil, nil, nil,
+	)
+	wrongChainTx.From = suite.signerAddress
+	_ = wrongChainTx.Sign(ethtypes.LatestSignerForChainID(wrongChainID), suite.signer)
+	wrongChainRlp, _ := rlp.EncodeToBytes(wrongChainTx.AsTransaction())
+
 	testCases := []struct {
 		name         string
 		registerMock func()
@@ -304,6 +324,23 @@ func (suite *BackendTestSuite) TestSendRawTransaction() {
 			"fail - no RLP encoded bytes",
 			func() {},
 			bz,
+			common.Hash{},
+			false,
+		},
+		{
+			"fail - fee exceeds configured cap",
+			func() {
+				suite.backend.allowUnprotectedTxs = true
+				suite.backend.cfg.JSONRPC.TxFeeCap = 1.0 // 1 ether cap
+			},
+			highFeeRlp,
+			common.Hash{},
+			false,
+		},
+		{
+			"fail - wrong chain ID",
+			func() { suite.backend.allowUnprotectedTxs = true },
+			wrongChainRlp,
 			common.Hash{},
 			false,
 		},